Multipath Support
With [1] Dracut now supports the ability to apply the default
configuration for Multipath devices. In order to support that we need to
make a few changes: notably we need to wait until multipath settles the
devices before we start using the device labels.

As a result, I needed to rework the excellent work done for
15coreos-copy-firstboot-network to be better-ordered. Like most of our
modules, we rely on simple devices with simple labels. The following was
done to support this goal:
- renamed 15coreos-copy-firstboot-network to 15coreos-firstboot-network
- allow the NetworkManager generators to run, but side-step the
  initqueue for firstboot only.

[1] https://github.com/dracutdevs/dracut/commit/b8a92b715677d52dbc2b27a710b9816fd8b9a63b

Signed-off-by: Ben Howard <ben.howard@redhat.com>